move helper to indexer module
