+++ src/indexer/mod.rs
@@ -26,7 +26,7 @@ use smallvec::SmallVec;
 
 pub use self::index_writer::IndexWriter;
 pub(crate) use self::json_term_writer::{
-    generate_term_from_json_writer, generate_terms_from_json_writer, JsonTermWriter,
+    generate_terms_from_json_writer, infer_fast_value_term, JsonTermWriter,
 };
 pub use self::log_merge_policy::LogMergePolicy;
 pub use self::merge_operation::MergeOperation;
